+++ Assets/Animations/Orc/RunFWD.anim
@@ -39951,10 +39951,24 @@ AnimationClip:
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
+  - time: 0.4
+    functionName: PlaySoundOnLocation
+    data: 
+    objectReferenceParameter: {fileID: 11400000, guid: 669e3af9fe0e44b4ab4dc8ae2387176b, type: 2}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
   - time: 0.6666667
     functionName: StompEffectL
     data: 
     objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
+  - time: 0.6666667
+    functionName: PlaySoundOnLocation
+    data: 
+    objectReferenceParameter: {fileID: 11400000, guid: 669e3af9fe0e44b4ab4dc8ae2387176b, type: 2}
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
