+++ Assets/Animations/Player/Mixamo/Landing.anim
@@ -25232,4 +25232,11 @@ AnimationClip:
   m_EulerEditorCurves: []
   m_HasGenericRootTransform: 0
   m_HasMotionFloatCurves: 0
-  m_Events: []
+  m_Events:
+  - time: 0
+    functionName: PlaySoundOnLocation
+    data: 
+    objectReferenceParameter: {fileID: 11400000, guid: 479403d503d686043bf2ba716eb86511, type: 2}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
