+++ Assets/Animations/Player/Mixamo/Turning-Left.anim
@@ -22408,4 +22408,11 @@ AnimationClip:
     script: {fileID: 0}
   m_HasGenericRootTransform: 0
   m_HasMotionFloatCurves: 0
-  m_Events: []
+  m_Events:
+  - time: 0.033333335
+    functionName: PlaySoundOnLocation
+    data: 
+    objectReferenceParameter: {fileID: 11400000, guid: 83af3a2da1f04b2459414a092d9bc40c, type: 2}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
