--- Assets/AudioData/FootStepPlayerGrass.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f6fb6f9bc3e121e48a029b598379655e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
