--- Assets/AudioData/LightSwingPlayer.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 784ac597430e72f439e69aea7fda77d6
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
