+++ Assets/AudioData/Orc/Orc-SpinAttack.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c125b484c3786114680c5e93e0e1e788
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
