+++ Assets/AudioData/Orc/Orc-SwingAttack.asset
@@ -1,0 +1,23 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 06a1fa782b440104587e3d08957860e1, type: 3}
+  m_Name: Orc-SwingAttack
+  m_EditorClassIdentifier: 
+  eventName: OrcStep
+  eventPath:
+    Guid:
+      Data1: -74848044
+      Data2: 1130497808
+      Data3: 941783206
+      Data4: -1235099136
+    Path: event:/SFX/Orc/SwingAttack
+  eventLocation: {fileID: 0}
