+++ Assets/AudioData/Orc/Orc-SwingAttack.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: de097c5e61174164a873e8847a82aa3f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
