+++ Assets/AudioData/Orc/Orc-Taunt.asset
@@ -1,0 +1,23 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 06a1fa782b440104587e3d08957860e1, type: 3}
+  m_Name: Orc-Taunt
+  m_EditorClassIdentifier: 
+  eventName: OrcStep
+  eventPath:
+    Guid:
+      Data1: 1604709656
+      Data2: 1161584349
+      Data3: 1442391479
+      Data4: -1495361935
+    Path: event:/SFX/Orc/Taunt
+  eventLocation: {fileID: 0}
