+++ Assets/AudioData/Orc/Orc-Taunt.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b588d4aa45780bf4c892d9fbbbbe0169
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
