+++ Assets/AudioData/Orc/Orc-WalkStep.asset
@@ -1,0 +1,23 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 06a1fa782b440104587e3d08957860e1, type: 3}
+  m_Name: Orc-WalkStep
+  m_EditorClassIdentifier: 
+  eventName: OrcStep
+  eventPath:
+    Guid:
+      Data1: 510406576
+      Data2: 1105312700
+      Data3: -972387912
+      Data4: 785256427
+    Path: event:/SFX/Orc/FootStep
+  eventLocation: {fileID: 0}
