+++ Assets/AudioData/Orc/Orc-WalkStep.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 669e3af9fe0e44b4ab4dc8ae2387176b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
