+++ Assets/AudioData/PlayerAction/Drifting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 83af3a2da1f04b2459414a092d9bc40c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
