+++ Assets/AudioData/PlayerAction/FootStepPlayerGrass.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f6fb6f9bc3e121e48a029b598379655e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
