+++ Assets/AudioData/PlayerAction/HeavySwingPlayer.asset
@@ -1,0 +1,23 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 06a1fa782b440104587e3d08957860e1, type: 3}
+  m_Name: HeavySwingPlayer
+  m_EditorClassIdentifier: 
+  eventName: HeavySwing
+  eventPath:
+    Guid:
+      Data1: 56497536
+      Data2: 1324866571
+      Data3: -1376926044
+      Data4: 634751631
+    Path: event:/SFX/Player/HeavySwing
+  eventLocation: {fileID: 0}
