+++ Assets/AudioData/PlayerAction/HeavySwingPlayer.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d0ef5250bf486234f9d790595671e869
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
