+++ Assets/AudioData/PlayerAction/JumpPlayer.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f47fec640252fd446a54efb6136e231c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
