+++ Assets/AudioData/PlayerAction/LandingPlayer.asset
@@ -1,0 +1,23 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 06a1fa782b440104587e3d08957860e1, type: 3}
+  m_Name: LandingPlayer
+  m_EditorClassIdentifier: 
+  eventName: Landing
+  eventPath:
+    Guid:
+      Data1: -1218693013
+      Data2: 1193220831
+      Data3: -74015061
+      Data4: -518671961
+    Path: event:/SFX/Player/Land
+  eventLocation: {fileID: 0}
