+++ Assets/AudioData/PlayerAction/LandingPlayer.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 479403d503d686043bf2ba716eb86511
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
