+++ Assets/_Scenes/_BGMManager.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3cf55b411ee1fe04cadaace0a4824401
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
