+++ Assets/_Scripts/Audio/MusicManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ed2a10d5a761f4246bf9641ae200b5fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
