+++ Assets/_Scripts/SharedMechanics/MovementAnimationEvents.cs
@@ -1,11 +1,8 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class MovementAnimationEvents : MonoBehaviour
 {
     [SerializeField] private ParticleSystem leftFeetStomp, rightFeetStomp;
-
 
     public void StompEffectL()
     {
