--- Assets/_Scripts/SharedMechanics/OneShotAudioData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 06a1fa782b440104587e3d08957860e1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
