+++ Assets/_UsedSprites/OrcPBR.prefab
@@ -1337,6 +1337,7 @@ GameObject:
   - component: {fileID: 232482936590703972}
   - component: {fileID: 2013657486944788246}
   - component: {fileID: 3141776805730351646}
+  - component: {fileID: 7955080600462142810}
   m_Layer: 10
   m_Name: OrcPBR
   m_TagString: Untagged
@@ -1395,7 +1396,7 @@ MonoBehaviour:
   weapon: {fileID: 174475150080888392}
   weaponTrail: {fileID: 2761847111763253225}
   weaponParticle: {fileID: 0}
-  impactParticle: {fileID: 0}
+  impactParticle: {fileID: 5011245444494639215}
 --- !u!114 &3141776805730351646
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1410,6 +1411,18 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   leftFeetStomp: {fileID: 9001871814736456491}
   rightFeetStomp: {fileID: 3490648664062393984}
+--- !u!114 &7955080600462142810
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 174287561557763434}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e0c63e63072fa2546bfcb971e4b5b1ec, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &174307002647609290
 GameObject:
   m_ObjectHideFlags: 0
@@ -2534,15 +2547,64 @@ PrefabInstance:
       propertyPath: m_Name
       value: SlashImpact
       objectReference: {fileID: 0}
+    - target: {fileID: 441730140364569802, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: playOnAwake
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 441730140364569802, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: moveWithTransform
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 457314784605242880, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: m_SortingLayer
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 457314784605242880, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: m_SortingLayerID
+      value: 1293318895
+      objectReference: {fileID: 0}
+    - target: {fileID: 458414992380764184, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: playOnAwake
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 458414992380764184, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: moveWithTransform
+      value: 1
+      objectReference: {fileID: 0}
     - target: {fileID: 5622100612644214210, guid: 154512af11dd90443ab5dcca27369470, type: 3}
       propertyPath: m_Intensity
       value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6122499776880690458, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: playOnAwake
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6122499776880690458, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: moveWithTransform
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 6681568392696145556, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: m_Enabled
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7416818520380549772, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: playOnAwake
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7416818520380549772, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+      propertyPath: moveWithTransform
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 154512af11dd90443ab5dcca27369470, type: 3}
 --- !u!4 &5141146811178083855 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 328233799518422648, guid: 154512af11dd90443ab5dcca27369470, type: 3}
+  m_PrefabInstance: {fileID: 4888405813069373559}
+  m_PrefabAsset: {fileID: 0}
+--- !u!198 &5011245444494639215 stripped
+ParticleSystem:
+  m_CorrespondingSourceObject: {fileID: 458414992380764184, guid: 154512af11dd90443ab5dcca27369470, type: 3}
   m_PrefabInstance: {fileID: 4888405813069373559}
   m_PrefabAsset: {fileID: 0}
 --- !u!1001 &6309860829756231406
@@ -2685,13 +2747,13 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 59db982d012a9b44ab04687e73027280, type: 3}
---- !u!198 &2761847111763253225 stripped
-ParticleSystem:
-  m_CorrespondingSourceObject: {fileID: 5547028746482850244, guid: 59db982d012a9b44ab04687e73027280, type: 3}
-  m_PrefabInstance: {fileID: 7687376359796756013}
-  m_PrefabAsset: {fileID: 0}
 --- !u!4 &2761847111155276487 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 5547028745999783146, guid: 59db982d012a9b44ab04687e73027280, type: 3}
+  m_PrefabInstance: {fileID: 7687376359796756013}
+  m_PrefabAsset: {fileID: 0}
+--- !u!198 &2761847111763253225 stripped
+ParticleSystem:
+  m_CorrespondingSourceObject: {fileID: 5547028746482850244, guid: 59db982d012a9b44ab04687e73027280, type: 3}
   m_PrefabInstance: {fileID: 7687376359796756013}
   m_PrefabAsset: {fileID: 0}
